refactor: moves method unzip to parent class
