+++ src/main/java/com/totalcross/DownloadManager.java
@@ -1,6 +1,13 @@
 package com.totalcross;
 
 import java.io.File;
+import java.io.IOException;
+import java.util.List;
+
+import org.codehaus.plexus.util.FileUtils;
+
+import net.lingala.zip4j.ZipFile;
+import net.lingala.zip4j.model.FileHeader;
 
 public abstract class DownloadManager {
    public static final boolean isWindows = System.getProperty("os.name").toLowerCase().startsWith("windows");
@@ -26,5 +33,28 @@ public abstract class DownloadManager {
 
    protected boolean verify(String subpath) {
       return new File(localRepositoryDir, subpath).exists();
+   }
+
+   protected abstract void rename(String from, String to) throws IOException;
+
+   public void unzip(String source, String dest) {
+      try {
+         ZipFile zipFile = new ZipFile(new File(localRepositoryDir, source));
+         if (!zipFile.getFile().exists())
+            return;
+         zipFile.extractAll(localRepositoryDir);
+         List<FileHeader> filesOnZip = zipFile.getFileHeaders();
+         String firstFileOnZip = filesOnZip.get(0).getFileName();
+         if (filesOnZip.get(0).isDirectory()) {
+            firstFileOnZip = firstFileOnZip.substring(0, firstFileOnZip.length() - 1);
+         }
+         rename(firstFileOnZip, dest);
+         FileUtils.deleteDirectory(new File(localRepositoryDir, firstFileOnZip));
+         FileUtils.deleteDirectory(new File(localRepositoryDir, source));
+
+      } catch (Exception e) {
+         e.printStackTrace();
+         System.exit(1);
+      }
    }
 }
